Split regtest integration tests by scenario
